Reverted some changes in console program, updates decoding code and documentation, library code separated to three units for better maintainability

+++ Source/ValueNodes/SII_Decode_ValueNode_00000019.pas
@@ -56,7 +56,7 @@ uses
 
 procedure TSIIBin_ValueNode_00000019.Initialize;
 var
-  Coef: Integer;
+  Coef: Int64;
 begin
 If FormatVersion = 2 then
   begin
